Fix module-level algolia init and unused var lint error
