+++ apps/asap-cli/src/scripts/algolia/set-settings.ts
@@ -13,7 +13,7 @@ export type SetAlgoliaSettings = {
 export const getIndexSchema = async ({
   algoliaAppId,
   algoliaCiApiKey,
-  indexName,
+  indexName: _indexName,
   appName,
 }: SetAlgoliaSettings): Promise<{
   path: string;
